# specs/004-app-logic-docs/contracts/openapi-excerpt.yaml
openapi: 3.0.3
info:
  title: Stock Verify API (Excerpt)
  version: "v-current"
  description: |
    Minimal OpenAPI excerpt for documentation purposes.
    This file is not intended to be complete or authoritative; validate against `/api/docs`.
servers:
  - url: http://localhost:8001
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/reports/query/preview:
    post:
      summary: Preview a report query
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Preview results
  /api/reports/snapshots:
    get:
      summary: List report snapshots
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Snapshot list
    post:
      summary: Create a report snapshot
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Snapshot created